rename query_class -> query_cls. remove namedtuple, use class vars.
return copy of query_cls on calls to query property.
We can't know if query_cls is immutable, so we must copy it.
We can use an immutablemultidict in the future if we remove query_cls from the
ctor. This would be my recommendation (let subclasses decide if a different
query class is appropriate.)